@@ -1,11 +1,7 @@
 import { useRef } from 'react';
-
 import useAutorun from './useAutorun';
 
-export default function useAutorunRef<T>(
-  callback: (ref: T) => void,
-  dependencies?: readonly any[]
-) {
+export default function useAutorunRef<T>(callback: (ref: T) => void, dependencies?: readonly any[]) {
   const ref = useRef<T | null>(null);
 
   useAutorun(() => {
