@@ -1,13 +1,13 @@
-import BaseDocument, { Head, Main, NextDocumentContext, NextScript } from 'next/document';
+import BaseDocument, { DocumentContext, Head, Main, NextScript } from 'next/document';
 import { ReactElement } from 'react';
 import { ServerStyleSheet } from 'styled-components';
 
-interface IProps {
+interface DocumentProps {
   styleTags: Array<ReactElement<HTMLStyleElement>>;
 }
 
-class Document extends BaseDocument<IProps> {
-  static getInitialProps({ renderPage }: NextDocumentContext) {
+class Document extends BaseDocument<DocumentProps> {
+  static async getInitialProps({ renderPage }: DocumentContext) {
     const sheet = new ServerStyleSheet();
     const page = renderPage(App => props => sheet.collectStyles(<App {...props} />));
     const styleTags = sheet.getStyleElement();
